refactor(BA-6630): move fair-share result types to the data layer
Relocate the fair-share factor-calculation and usage-bucket-aggregation result
types (FairShareFactorCalculationResult, UsageBucketAggregationResult and their
nested factor/key/delta types) from the sokovan calculator/aggregator modules
into data/fair_share, cutting the remaining repositories/fair_share and
repositories/resource_usage_history back-edges into the sokovan package.
