+++ src/ai/backend/manager/data/fair_share/__init__.py
@@ -1,20 +1,30 @@
 from .types import (
+    BucketDelta,
+    DomainFactorResult,
     DomainFairShareData,
     DomainFairShareSearchResult,
+    DomainUsageBucketKey,
     FairShareCalculationContext,
     FairShareCalculationSnapshot,
     FairShareData,
+    FairShareFactorCalculationResult,
     FairShareMetadata,
     FairSharesByLevel,
     FairShareSpec,
+    ProjectFactorResult,
     ProjectFairShareData,
     ProjectFairShareSearchResult,
+    ProjectUsageBucketKey,
     ProjectUserIds,
     RawUsageBucketsByLevel,
+    UsageBucketAggregationResult,
+    UserFactorResult,
     UserFairShareData,
     UserFairShareFactors,
     UserFairShareSearchResult,
     UserProjectKey,
+    UserSchedulingRank,
+    UserUsageBucketKey,
 )
 
 __all__ = (
@@ -39,4 +49,16 @@ __all__ = (
     # User-level
     "UserFairShareData",
     "UserFairShareSearchResult",
+    # Factor calculation results
+    "DomainFactorResult",
+    "ProjectFactorResult",
+    "UserFactorResult",
+    "UserSchedulingRank",
+    "FairShareFactorCalculationResult",
+    # Usage bucket aggregation results
+    "UserUsageBucketKey",
+    "ProjectUsageBucketKey",
+    "DomainUsageBucketKey",
+    "BucketDelta",
+    "UsageBucketAggregationResult",
 )
